+++ src/pdfTextExtractor/dataExtractor.js
