+++ src/pdfTextExtractor/pdfText.js
